Patch homework permissions + add logs

+++ modules/course/course-service/src/main/java/com/weprode/facile/course/service/impl/HomeworkLocalServiceImpl.java
@@ -467,12 +467,14 @@ public class HomeworkLocalServiceImpl extends HomeworkLocalServiceBaseImpl {
 
 		Homework homework = getHomework(homeworkId);
 		Folder courseFolder = FolderUtilsLocalServiceUtil.getGroupCourseFolder(homework.getCourseId());
+		PermissionUtilsLocalServiceUtil.addDefaultPermissionsFolder(courseFolder);
 
 		Folder homeworkFolder = null;
 		try {
 			homeworkFolder = FolderUtilsLocalServiceUtil.getFolderByName(courseFolder, String.valueOf(homeworkId));
 		} catch (NoSuchFolderException e) {
 			if (doCreate) {
+				logger.info("Creating folder for homework " + homeworkId);
 				homeworkFolder = DLAppServiceUtil.addFolder(
 						UUID.randomUUID().toString(),
 						courseFolder.getGroupId(),
@@ -481,6 +483,7 @@ public class HomeworkLocalServiceImpl extends HomeworkLocalServiceBaseImpl {
 						"Dossier du devoir " + homeworkId,
 						new ServiceContext()
 				);
+				logger.info("Created drop folder for homework " + homeworkId);
 			}
 		} catch (Exception e) {
 			logger.error("Error when fetching folder for homeworkId " + homeworkId, e);
@@ -495,24 +498,29 @@ public class HomeworkLocalServiceImpl extends HomeworkLocalServiceBaseImpl {
 		return homeworkFolder;
 	}
 
-	public Folder getHomeworkDropFolder(long homeworkId) throws PortalException, SystemException {
-
-		Folder homeworkFolder = getHomeworkFolder(homeworkId, true);
+	public Folder getHomeworkDropFolder(long homeworkId, boolean doCreate) throws PortalException, SystemException {
 
 		Folder homeworkDropFolder = null;
-		try {
-			homeworkDropFolder = FolderUtilsLocalServiceUtil.getFolderByName(homeworkFolder, CourseConstants.DROP_FOLDER);
-		} catch (NoSuchFolderException e) {
-			homeworkDropFolder = DLAppServiceUtil.addFolder(
-					UUID.randomUUID().toString(),
-					homeworkFolder.getGroupId(),
-					homeworkFolder.getFolderId(),
-					CourseConstants.DROP_FOLDER,
-					"",
-					new ServiceContext()
-			);
-		} catch (Exception e) {
-			logger.error("Error when fetching folder for homeworkId " + homeworkId, e);
+		Folder homeworkFolder = getHomeworkFolder(homeworkId, doCreate);
+
+		if (homeworkFolder != null) {
+			try {
+				homeworkDropFolder = FolderUtilsLocalServiceUtil.getFolderByName(homeworkFolder, CourseConstants.DROP_FOLDER);
+			} catch (NoSuchFolderException e) {
+				if (doCreate) {
+					logger.info("Creating drop folder for homework " + homeworkId);
+					homeworkDropFolder = DLAppServiceUtil.addFolder(
+							UUID.randomUUID().toString(),
+							homeworkFolder.getGroupId(),
+							homeworkFolder.getFolderId(),
+							CourseConstants.DROP_FOLDER,
+							"",
+							new ServiceContext()
+					);
+					logger.info("Created drop folder for homework " + homeworkId);
+				}
+			}
+
 		}
 
 		return homeworkDropFolder;
@@ -547,8 +555,16 @@ public class HomeworkLocalServiceImpl extends HomeworkLocalServiceBaseImpl {
 			StudentHomeworkLocalServiceUtil.removeHomework(homeworkId);
 
 			// Delete drop folder
-			Folder homeworkDropFolder = getHomeworkDropFolder(homeworkId);
-			DLAppServiceUtil.deleteFolder(homeworkDropFolder.getFolderId());
+			Folder homeworkDropFolder = getHomeworkDropFolder(homeworkId, false);
+			if (homeworkDropFolder != null) {
+				DLAppServiceUtil.deleteFolder(homeworkDropFolder.getFolderId());
+			}
+
+			// Delete homework folder
+			Folder homeworkFolder = getHomeworkFolder(homeworkId, false);
+			if (homeworkFolder != null) {
+				DLAppServiceUtil.deleteFolder(homeworkFolder.getFolderId());
+			}
 
 			// Remove the homework itself
 			deleteHomework(homeworkId);
